+++ CHANGES.md
@@ -1,4 +1,4 @@
-= 3.1.4 (2025//) =
+= 3.1.4 (2025/07/21) =
 * Added new field option "toolbaroptions". Works like "options", but only for the toolbar (in case you want the regular fields to be just the supplied text)
 * dependsOn can now refer to any field, not just another dropdown. So you can create a dropdown depending on a text field too (as an example)
 * Added trigger "toolbarsearch", so you can hook into this if need be
@@ -6,6 +6,7 @@
 * flatpickr also supported as datepicker (also returns Y-m-d to the backend)
 * UI datepicker now also expects/returns in Y-m-d format, so all datepickers are now identical
 * support datetime-local too, with respect for displayFormat
+* renamed field options displayFormat to dateFormat and displayLocale to dateLocale
 
 = 3.1.3 (2025/07/15) =
 * the toolbarsearch extension is now in the main plugin